third_party/WebKit/LayoutTests/css3/tab-size-expected.txt

Test tab-size measurements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pxTabLength is pxExpected
PASS emTabLength is emExpected
PASS intTabLength is intExpected
PASS pxTabLength is pxExpected
PASS emTabLength is emExpected
PASS intTabLength is intExpected
PASS successfullyParsed is true

TEST COMPLETE

	leading text, tab-size in px.
	leading text, tab-size in em.
	leading text, tab-size in spaces.
leading text	trailing text, tab-size in px.
leading text	trailing text, tab-size in em.
leading text	trailing text, tab-size in spaces.
Some spaces... '         ' ... for size reference.
'leading text', for size reference.
